+++ Runtime/Extensions/AudioSourceExtension.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+
+namespace ActionCode.Audio
+{
+    /// <summary>
+    /// Extension class for <see cref="AudioSource"/>.
+    /// </summary>
+    public static class AudioSourceExtension
+    {
+        /// <summary>
+        /// Sets the Audio source to play as a 2D sound.
+        /// </summary>
+        /// <param name="source"></param>
+        public static void SetSpatialBlendTo2D(this AudioSource source) => source.spatialBlend = 0F;
+
+        /// <summary>
+        /// Sets the Audio source to play as a 3D sound.
+        /// </summary>
+        /// <param name="source"></param>
+        public static void SetSpatialBlendTo3D(this AudioSource source) => source.spatialBlend = 1F;
+    }
+}