+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Remove DisallowMultipleComponent attribute from AudioSourceDictionary
 - Remove UI Components from AudioGroup
+- If possible, draws AudioGroup.volumeParamName as a Popup field
 
 ### Removed
 - Text Mesh Pro package dependency