+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AudioSourceAudioGroup component
 - AudioSourceExtension
 - AudioGroupSettings
+- AmbientEffectsVolumeGroupSettings asset
+- BackgroundVolumeGroupSettings asset
+- SoundEffectsVolumeGroupSettings asset
+- VoiceEffectsVolumeGroupSettings asset
 
 ### Changed
 - Remove DisallowMultipleComponent attribute from AudioSourceDictionary