+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - Text Mesh Pro package dependency
+- AudioGroupManager component
 
 ## [2.0.0] - 2023-01-24