+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AudioGroupSettings ScriptableObject class
 - AmbientEffects, Background, SoundEffects and VoiceEffects Group Settings assets
 - No Effects Snapshot
-- AudioSettings Scriptable Object
+- AudioSettings Scriptable Object and asset
 
 ### Changed
 - Remove DisallowMultipleComponent attribute from AudioSourceDictionary